+++ htdocs/js/globales.js
@@ -1029,6 +1029,14 @@ const settingsConnectionFields = {
         "unit": "kB",
         "class": ["alwaysEnabled"]
     },
+    "mpdStringnormalization": {
+        "defaultValue": defaults["MYMPD_MPD_STRINGNORMALIZATION"],
+        "inputType": "checkbox",
+        "title": "String normalization",
+        "form": "modalConnectionAdvFrm2",
+        "help": "helpConnectionStringnormalization",
+        "class": ["alwaysEnabled", "featStringnormalization"]
+    },
     "stickerdbMpdHost": {
         "defaultValue": defaults["MYMPD_MPD_HOST"],
         "inputType": "text",